c syntax - use sub calls to clang to parse qualified types
This is probably not the best approach, but it unlocks a lot of AST
coverage: clang's json output doesn't refine the type of variable
declaration, typecast etc. but it does for typedef and type alias. So we
generate a fake type declaration and call clang again on it to re-parse
the type and get a fine grain definition of it.