Proper type collection
Instead of trying to reparse types to extract their full tree, use a
Clang Plugin that performs a first tree transversal to gather the full
type of every node, then a second regular transversal to dump the AST.
The link between the two dumps is made through expression unique id,
which work because the two passes work on the same AST.

The plugin approach gives access to the full AST representation, so this
is much more reliable than other approaches. For completeness, let's
mention that the clang Python binding have been tried but its output is
(much more) incomplete than the JSON output. It's also not possible to
just extend the JSON dumper because it has not been designed upstream
to be customizable through subclassing.

The compilation of a clang plugin assumes the clang header are available
on the system.