c support - asm statement
The JSON output misses the ASM String literal. This requires an extra
processing in the clang plugin, but we have enough machinery to do so.

The support is probably not complete as this statement is rather
complex, but we at least support basic input / output operands,
registers clobber and labels.