c syntax - support __typeof__ keyword
This is preliminary work for decltype support, but we hit (again) the
wall of expression-dependent types. We currently by-pass the issue by
using the expression representation instead of it's AST, but this is a
very limited approach in terms of quality of the AST representation.

Yet it makes it possible to progress, so that still looks like a decent
decision.